requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
-e git+https://github.com/theovoss/Chess@c0b1c8838a8fef46a3e3035e8285e14cebe426c4#egg=chess
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
dj-database-url==0.4.2
django==3.1
django-allauth==0.42.0
django-cors-headers==3.4.0
django-crispy-forms==1.9.2
django-grappelli==2.14.2
django-redis-cache==2.1.1
djangorestframework==3.11.1
drf-nested-routers==0.91
drf-yasg==1.17.1
faker==4.1.1
gunicorn==20.0.4
idna==2.10
inflection==0.5.0
itypes==1.2.0
jinja2==2.11.2
markupsafe==1.1.1
minilog==1.6
oauthlib==3.1.0
packaging==20.4
psycopg2-binary==2.8.5
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0; platform_python_implementation == "CPython" and python_version < "3.9"
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.25.10
whitenoise==5.2.0
